pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contract-simulator"
version = "0.1.0"
description = "Contract Simulation & Stress-Tester — simulate how contract clauses play out under various scenarios"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Noam Raz" },
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.42.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "python-multipart>=0.0.18",
    "streamlit>=1.41.0",
    "requests>=2.32.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
